+++ extensions/downloads-manager/src/open-downloads.tsx
@@ -1,0 +1,7 @@
+import { popToRoot, showInFinder } from "@raycast/api";
+import { downloadsDir } from "./utils";
+
+export default async function main() {
+  await showInFinder(downloadsDir);
+  await popToRoot();
+}