+++ extensions/downloads-manager/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Downloads Manager Changelog
 
+## [Add new commands] - 2022-10-08
+- adds `Copy Latest Download` command
+- adds `Open Downloads Folder` command
+
 ## [Initial Version] - 2022-08-22